+++ packages/browser/lib/utils/errors.ts
@@ -23,3 +23,5 @@ export class ExceededMaxFileSizeError extends EvervaultError {}
 export class DecryptError extends EvervaultError {}
 
 export class FormError extends EvervaultError {}
+
+export class DataRolesNotSupportedError extends EvervaultError {}