+++ packages/ui-components/src/Card/index.tsx
@@ -1,19 +1,22 @@
+import {
+  validateNumber,
+  validateCVC,
+  validateExpiry,
+} from "@evervault/card-validator";
 import { useEvervault } from "@evervault/react";
-import cardValidator from "card-validator";
 import { useEffect, useLayoutEffect, useMemo, useRef } from "react";
+import { useForm, useTranslations } from "shared";
 import { Error } from "../Common/Error";
 import { Field } from "../Common/Field";
 import { resize } from "../utilities/resize";
-import { useForm } from "../utilities/useForm";
 import { useMessaging } from "../utilities/useMessaging";
-import { useTranslations } from "../utilities/useTranslations";
 import { CardCVC } from "./CardCVC";
 import { CardExpiry } from "./CardExpiry";
 import { CardHolder } from "./CardHolder";
 import { CardNumber } from "./CardNumber";
 import { DEFAULT_TRANSLATIONS } from "./translations";
 import { useCardReader } from "./useCardReader";
-import { changePayload, isCVCValid, swipePayload } from "./utilities";
+import { changePayload, isAcceptedBrand, swipePayload } from "./utilities";
 import type { CardForm, CardConfig } from "./types";
 import type { CardFrameClientMessages, CardFrameHostMessages } from "types";
 
@@ -25,6 +28,8 @@ export function Card({ config }: { config: CardConfig }) {
   >();
   const ev = useEvervault();
   const { t } = useTranslations(DEFAULT_TRANSLATIONS, config?.translations);
+
+  const { acceptedBrands } = config;
 
   const fields = useMemo(() => {
     let result = config.fields ?? ["number", "expiry", "cvc"];
@@ -57,9 +62,13 @@ export function Card({ config }: { config: CardConfig }) {
       number: (values) => {
         if (!fields.includes("number")) return undefined;
 
-        const cardValidation = cardValidator.number(values.number);
+        const cardValidation = validateNumber(values.number);
         if (!cardValidation.isValid) {
           return "invalid";
+        }
+
+        if (!isAcceptedBrand(acceptedBrands, cardValidation)) {
+          return "unsupportedBrand";
         }
 
         return undefined;
@@ -67,8 +76,8 @@ export function Card({ config }: { config: CardConfig }) {
       expiry: (values) => {
         if (!fields.includes("expiry")) return undefined;
 
-        const expiry = cardValidator.expirationDate(values.expiry);
-        if (!expiry.isValid) {
+        const expiryValidation = validateExpiry(values.expiry);
+        if (!expiryValidation.isValid) {
           return "invalid";
         }
 
@@ -77,9 +86,8 @@ export function Card({ config }: { config: CardConfig }) {
       cvc: (values) => {
         if (!fields.includes("cvc")) return undefined;
 
-        const cardValidation = cardValidator.number(values.number);
-        const validCVC = isCVCValid(values.cvc, cardValidation.card?.type);
-        if (!validCVC) {
+        const cvcValidation = validateCVC(values.cvc, values.number);
+        if (!cvcValidation.isValid) {
           return "invalid";
         }