+++ e2e-apps/browser/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- @evervault/browser@2.23.2
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [8fe5c04]
+  - @evervault/browser@2.23.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [a96e24b]
+  - @evervault/browser@2.23.0
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [43d5607]
   - @evervault/browser@2.22.1