+++ e2e-apps/decrypt-backend/CHANGELOG.md
@@ -149,3 +149,21 @@
 ### Patch Changes
 
 - @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null
+
+## null
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@null