+++ examples/forms/CHANGELOG.md
@@ -1,5 +1,25 @@
 # example-forms
 
+## 0.0.6
+
+### Patch Changes
+
+- @evervault/browser@2.23.2
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [8fe5c04]
+  - @evervault/browser@2.23.1
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [a96e24b]
+  - @evervault/browser@2.23.0
+
 ## 0.0.3
 
 ### Patch Changes