+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @evervault/browser
 
+## 2.23.2
+
+### Patch Changes
+
+- Updated dependencies [1ff4434]
+  - themes@0.0.1
+  - types@0.0.1
+
+## 2.23.1
+
+### Patch Changes
+
+- 8fe5c04: Removed Data Role Encryption for Files
+
+## 2.23.0
+
+### Minor Changes
+
+- a96e24b: Add acceptedBrands option
+
 ## 2.22.1
 
 ### Patch Changes